pyproject.toml
[project]
name = "shared-auth-lib"
version = "0.2.0"
description = "Shared authorization library for ThinkRealty microservices"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.104.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.25.0",
    "starlette>=0.27.0",
    "tr-shared-lib[http,logging]>=0.4.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.25.0",
    "ruff>=0.4.0",
    "freezegun>=1.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["shared_auth_lib"]

[tool.uv.sources]
tr-shared-lib = { path = "../tr-shared-lib", editable = true }

[tool.ruff]
line-length = 85
target-version = "py313"

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=shared_auth_lib --cov-report=term-missing"

[tool.coverage.run]
source = ["shared_auth_lib"]
omit = ["*/tests/*", "*/__init__.py"]

[tool.coverage.report]
fail_under = 80
show_missing = true
skip_covered = false
